ANDROID: HACK: Makefile: whitelist CFLAGS to reenable Werror
This is a two pronged approached, based on whether offending files are
"in tree" vs "out of tree."

For "in tree" files that produce warnings, these warnings are
whitelisted in the top level Makefile. We will work with upstream to
either fix these warnings properly or silence them when building with
Clang.

For "out of tree" files, the most specific directory Makefile is used to
silence this warning for the sub directory.  We will work with the
appropriate vendors to fix these warnings.

With these existing warnings silenced, we can re-enable CONFIG_WERROR.
The method used for "out of tree" files will _not_ catch new instances
of the same error in the same subdirectory, but the idea is that no new
classes of warnings will be added in the future. As we fix existing
warnings, the whitelists will be pruned back (until they no longer
exist).

It's unlikely this patch will be reverted at once. Rather, as we fix
classes of warnings, the whitelists should be pruned back as applicable
until there are no more `TODO: remove me b/62057517`.

Bug: 37513383
Bug: 62057517
Bug: 62058353

Change-Id: I583552512a28eb8867a3f3e848be4af2becdbc8e
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: SagarMakhar <sagarmakhar@gmail.com>